Use jsonrpc2 for communication between acme-lsp and L
The plumber has been replaced with jsonrpc2 over a unix domain socket at
`$(namespace)/acme-lsp.rpc`. The RPC messages are currently similar to a
plumb message, but the plan is to change the messages to be a subset of
the LSP protocol.

We still depends on the plumber for `L def` and `L type`.

Update #14