Add first completion target with multiple matches
When a completion target has multiple matches, such as when editing Go code that uses generated Protobuf packages that have the same method name at multiple levels of hierarchy, insert the first match always, and display the rest in an Acme window called `/LSP/Completions` similar to `/LSP/Diagnostics`.

This behaviour can be triggered by passing `-E` to `L comp` instead of `-e`. The existing behaviour for `-e` is preserved as is.