acme-lsp: use standard acme range syntax, relative paths in L output
/usr/local/plan9/plumb/fileaddr defines twocolonaddr as line:char or line.char,
but line.char seems clearer given the colon is already being used to separate
the file name.

More importantly, plumb/basic uses twocolonaddr,twocolonaddr not
twocolonaddr-twocolonaddr, so use comma instead of dash to separate
the elements.

Also, print relative path names when that's shorter than the absolute path.
This is the same heuristic the go command uses for its own error messages,
so this makes L consistent with the go command and also just makes
the output easier to read.