internal/lsp: fix CompatibleCodeActions
CompatibleCodeActions was not expecting map[string]any,
but that's the type passed by the call in acmelsp.CodeActionAndFormat,
because InitializeResult.Capabilities is in fact a map[string]any.

The result was that CodeActionAndFormat never thought it could
use source.organizeImports, so Put was not adding new Go imports.

Now Put can add imports (again?).