Add "IgnorePattern" to file name handlers to explicitly ignore some paths
Some setups require running the same language server with slightly different
configurations for different paths. One example would be "I need `gopls` with
the Bazel package driver for exactly one project, but I want a regular `gopls`
for all other Go files".

This is a bit cumbersome to express with just the positive match pattern, so
this adds an explicit "ignore all files that match this pattern, even if they
match `Pattern`" option to `FilenameHandlers`.