Fix completion, code formatting, and hover help for dartls
The Dart LSP insists on getting the invocation kind for completions,
and on the diagnostics for the "Format" code action to be empty,
but _not_ `null`.

Dart's LS sends hover help responses as the deprecated
protocol.MarkedString instead of protocol.MarkupContent.

This requires [a tweak to the generated LSP protocol code](https://github.com/9fans/go-lsp-internal/pull/4) as well.