cmd/L: add wss sub-command for workspace symbol query
Add the wss sub-command which prints results for a symbol query as resolved by
all configured LSPs. This enables functionality like search or resolution of
symbols from plumber. As an example, plumber can match to java stack traces and
plumb to the source file resolved by L sym and line in the stack trace.

Fixes #90